refactor: extract context hook utilities and shared sentinel management
- Create src/cli/commands/context.ts with addContextHook/removeContextHook/hasContextHook
  extracted from init.ts (SRP: each hook feature in its own module)
- Create src/cli/utils/sentinel.ts with manageSentinel() to eliminate 6 copies
  of the identical enable/disable sentinel pattern across init.ts, memory.ts, learn.ts
- Update init.ts: import from context.ts and sentinel.ts; 4 sentinel blocks collapsed
  to a single manageSentinel call each; re-export context utilities for backward compat
- Update memory.ts: use manageSentinel; add best-effort queue drain on --disable
  (removes .pending-turns.jsonl and .pending-turns.processing orphans)
- Update learn.ts: use manageSentinel for --enable and --disable sentinel management
- Update uninstall.ts: import removeContextHook from context.ts (canonical source)
