+++ src/cli/commands/init.ts
@@ -23,7 +23,7 @@ import { detectPlatform, detectShell, getProfilePath, getSafeDeleteInfo, hasSafe
 import { generateSafeDeleteBlock, installToProfile, removeFromProfile, getInstalledVersion, SAFE_DELETE_BLOCK_VERSION } from '../utils/safe-delete-install.js';
 import { addAmbientHook, removeAmbientHook } from './ambient.js';
 import { addMemoryHooks, removeMemoryHooks } from './memory.js';
-import type { Settings, HookMatcher } from '../utils/hooks.js';
+// Settings/HookMatcher types used by hook utilities — each in their own module
 import { addLearningHook, removeLearningHook } from './learn.js';
 import { addDecisionsHook, removeDecisionsHook } from './decisions.js';
 import { addHudStatusLine, removeHudStatusLine } from './hud.js';
@@ -31,6 +31,8 @@ import { addKnowledgeHook, removeKnowledgeHook } from './knowledge/index.js';
 import { loadConfig as loadHudConfig, saveConfig as saveHudConfig } from '../hud/config.js';
 import { readManifest, writeManifest, resolvePluginList, detectUpgrade } from '../utils/manifest.js';
 import { getDefaultFlags, applyFlags, stripFlags, applyViewMode, stripViewMode, FLAG_REGISTRY, ViewMode, VIEW_MODES } from '../utils/flags.js';
+import { addContextHook, removeContextHook, hasContextHook } from './context.js';
+import { manageSentinel } from '../utils/sentinel.js';
 import * as os from 'os';
 
 // Re-export pure functions for tests (canonical source is post-install.ts)
@@ -100,90 +102,7 @@ export function classifySafeDeleteState(
   return 'missing';
 }
 
-// ─── Context hook utilities ────────────────────────────────────────────────
-//
-// The session-start-context hook is always-on (registered unconditionally by
-// init, removed by uninstall). It has internal sentinel awareness per feature.
-
-const CONTEXT_HOOK_MARKER = 'session-start-context';
-
-/**
- * Add the session-start-context hook to SessionStart in settings JSON.
- * Idempotent — returns unchanged JSON if hook already present.
- */
-export function addContextHook(settingsJson: string, devflowDir: string): string {
-  if (hasContextHook(settingsJson)) {
-    return settingsJson;
-  }
-
-  const settings: Settings = JSON.parse(settingsJson);
-
-  if (!settings.hooks) {
-    settings.hooks = {};
-  }
-
-  const hookCommand = path.join(devflowDir, 'scripts', 'hooks', 'run-hook') + ` ${CONTEXT_HOOK_MARKER}`;
-  const newEntry: HookMatcher = {
-    hooks: [
-      {
-        type: 'command',
-        command: hookCommand,
-        timeout: 10,
-      },
-    ],
-  };
-
-  if (!settings.hooks.SessionStart) {
-    settings.hooks.SessionStart = [];
-  }
-
-  settings.hooks.SessionStart.push(newEntry);
-
-  return JSON.stringify(settings, null, 2) + '\n';
-}
-
-/**
- * Remove the session-start-context hook from settings JSON.
- * Idempotent — returns unchanged JSON if hook not present.
- * Preserves all other SessionStart hooks.
- */
-export function removeContextHook(settingsJson: string): string {
-  const settings: Settings = JSON.parse(settingsJson);
-
-  if (!settings.hooks?.SessionStart) {
-    return settingsJson;
-  }
-
-  const before = settings.hooks.SessionStart.length;
-  settings.hooks.SessionStart = settings.hooks.SessionStart.filter(
-    (matcher) => !matcher.hooks.some((h) => h.command.includes(CONTEXT_HOOK_MARKER)),
-  );
-
-  if (settings.hooks.SessionStart.length === before) {
-    return settingsJson;
-  }
-
-  if (settings.hooks.SessionStart.length === 0) {
-    delete settings.hooks.SessionStart;
-  }
-
-  if (Object.keys(settings.hooks).length === 0) {
-    delete settings.hooks;
-  }
-
-  return JSON.stringify(settings, null, 2) + '\n';
-}
-
-/**
- * Check if the session-start-context hook is present in settings JSON.
- * Accepts either a JSON string or a parsed Settings object.
- */
-export function hasContextHook(input: string | Settings): boolean {
-  const settings: Settings = typeof input === 'string' ? JSON.parse(input) : input;
-  return settings.hooks?.SessionStart?.some(
-    (matcher) => matcher.hooks.some((h) => h.command.includes(CONTEXT_HOOK_MARKER)),
-  ) ?? false;
-}
+export { addContextHook, removeContextHook, hasContextHook };
 
 /**
  * Parse a comma-separated plugin selection string into normalized plugin names.
@@ -1223,49 +1142,12 @@ export const initCommand = new Command('init')
       }
     }
 
-    // Manage .disabled sentinel based on knowledgeEnabled state
+    // Manage runtime-disable sentinels for each feature
     if (gitRoot) {
-      const disabledPath = path.join(gitRoot, '.features', '.disabled');
-      if (knowledgeEnabled) {
-        try { await fs.unlink(disabledPath); } catch { /* doesn't exist */ }
-      } else {
-        await fs.mkdir(path.join(gitRoot, '.features'), { recursive: true });
-        await fs.writeFile(disabledPath, '', 'utf-8');
-      }
-    }
-
-    // Manage .disabled sentinel for decisions based on decisionsEnabled state
-    if (gitRoot) {
-      const decisionsSentinelDir = path.join(gitRoot, '.memory', 'decisions');
-      const decisionsSentinel = path.join(decisionsSentinelDir, '.disabled');
-      if (decisionsEnabled) {
-        try { await fs.unlink(decisionsSentinel); } catch { /* doesn't exist */ }
-      } else {
-        await fs.mkdir(decisionsSentinelDir, { recursive: true });
-        await fs.writeFile(decisionsSentinel, '', 'utf-8');
-      }
-    }
-
-    // Manage .working-memory-disabled sentinel based on memoryEnabled state
-    if (gitRoot) {
-      const memorySentinel = path.join(gitRoot, '.memory', '.working-memory-disabled');
-      if (memoryEnabled) {
-        try { await fs.unlink(memorySentinel); } catch { /* doesn't exist */ }
-      } else {
-        await fs.mkdir(path.join(gitRoot, '.memory'), { recursive: true });
-        await fs.writeFile(memorySentinel, '', 'utf-8');
-      }
-    }
-
-    // Manage .learning-disabled sentinel based on learnEnabled state
-    if (gitRoot) {
-      const learnSentinel = path.join(gitRoot, '.memory', '.learning-disabled');
-      if (learnEnabled) {
-        try { await fs.unlink(learnSentinel); } catch { /* doesn't exist */ }
-      } else {
-        await fs.mkdir(path.join(gitRoot, '.memory'), { recursive: true });
-        await fs.writeFile(learnSentinel, '', 'utf-8');
-      }
+      await manageSentinel(gitRoot, path.join(gitRoot, '.features', '.disabled'), knowledgeEnabled);
+      await manageSentinel(gitRoot, path.join(gitRoot, '.memory', 'decisions', '.disabled'), decisionsEnabled);
+      await manageSentinel(gitRoot, path.join(gitRoot, '.memory', '.working-memory-disabled'), memoryEnabled);
+      await manageSentinel(gitRoot, path.join(gitRoot, '.memory', '.learning-disabled'), learnEnabled);
     }
 
     // Configure HUD
